AO3-6547 Extract WorkRecipients concern from Work model
